+++ ir/opt/scalar_replace.c
@@ -508,7 +508,7 @@ static unsigned allocate_value_numbers(pset *sels, ir_entity *ent, unsigned vnum
 			SET_VNUM(sel, key->vnum);
 			DB((dbg, SET_LEVEL_3, "  %+F represents value %u\n", sel, key->vnum));
 
-			ARR_EXTO(ir_mode *, *modes, (int)((key->vnum + 15) & ~15));
+			ARR_EXTO(ir_mode *, *modes, (key->vnum + 15) & ~15);
 
 			(*modes)[key->vnum] = get_type_mode(get_entity_type(get_Sel_entity(sel)));
